# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API do boilerplate GOTH Stack (Go, Templ, HTMX).",
        "title": "GOTH Stack API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/",
    "paths": {
        "/webhooks/{source}": {
            "post": {
                "description": "Recebe um payload de webhook, persiste no banco e enfileira um job.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "webhooks"
                ],
                "summary": "Receber Webhook",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Fonte do webhook (ex: stripe)",
                        "name": "source",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Payload do webhook",
                        "name": "payload",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/webhook.webhookInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "webhook.webhookInput": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "id": {
                    "type": "string"
                }
            }
        }
    }
}